refactor: 메소드 분리 및 파일 네이밍 변경
