+++ .gitignore
@@ -105,4 +105,5 @@ dist
 # API_KEY
 src/api.js
 
-dummyData
+#dummmyData
+dummyData.js
