--- src/js/util/dom.js
@@ -1,2 +1,0 @@
-export const $ = selector => document.querySelector(selector);
-export const $$ = selector => document.querySelectorAll(selector);
